backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
dj-database-url==2.2.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
django-cors-headers==4.4.0
gunicorn==22.0.0
python-dotenv==1.0.1
whitenoise==6.7.0
hypothesis==6.112.1
pytest==8.3.3
pytest-django==4.9.0
